pyproject.toml
[project]
name = "hydra-sqlalchemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.13.3",
    "hydra-core>=1.3.2",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = []
